requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
pillow==10.2.0
opencv-python==4.9.0.80
matplotlib==3.8.3
